// src/Interfaces/IMega.ts
export default interface IMega {
    _id: string
    number1: string
    number2: string
    number3: string
    number4: string
    number5: string
    number6: string
    jackpot: string
    jackpotWinner: string | null
    match5: string
    match4: string
    match5Winner: string
    match4Winner: string
    match3: string
    match3Winner: string
    dayPrize: string
    idKy: string
}