// src/Interfaces/IMax3DPro.ts
export default interface IMax3DPro {
    dayPrize: string
    specialPrize1: string
    specialPrize2: string
    subSpecialPrize1: string
    subSpecialPrize2: string
    firstPrize1: string
    firstPrize2: string
    firstPrize3: string
    firstPrize4: string
    secondPrize1: string
    secondPrize2: string
    secondPrize3: string
    secondPrize4: string
    secondPrize5: string
    secondPrize6: string
    thirdPrize1: string
    thirdPrize2: string
    thirdPrize3: string
    thirdPrize4: string
    thirdPrize5: string
    thirdPrize6: string
    thirdPrize7: string
    thirdPrize8: string
    specialPrizeWinners: string
    subSpecialPrizeWinners: string
    firstPrizeWinners: string
    secondPrizeWinners: string
    thirdPrizeWinners: string
    idKy: string
    _id: string
}