// src/Interfaces/IPower655.ts
export default interface IPower655 {
    dayPrize: string
    dayPrizeNext: string
    number1: string
    number2: string
    number3: string
    number4: string
    number5: string
    number6: string
    number7: string
    jackpot1: string
    jackpot2: string
    match3: string
    match4: string
    match5: string
    jackpotWinner: string
    jackpot2Winner: string
    match3Winner: string
    match4Winner: string
    match5Winner: string
    idKy: string
}